+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/FileSystemsTest.java
@@ -1942,12 +1942,12 @@ public class FileSystemsTest {
                 Assert.fail("Should not reach here.");
             } catch (Exception e) {
                 if (isInternal) {
-                    Assert.assertFalse(interop.hasHostObject(e) && interop.isException(e));
+                    Assert.assertFalse(interop.isHostObject(e) && interop.isException(e));
                     Assert.assertTrue(e instanceof RuntimeException);
                 } else {
-                    Assert.assertTrue(interop.hasHostObject(e) && interop.isException(e));
+                    Assert.assertTrue(interop.isHostObject(e) && interop.isException(e));
                     if (weakEncapsulation) {
-                        Assert.assertTrue(interop.getHostObject(e) instanceof NullPointerException);
+                        Assert.assertTrue(interop.asHostObject(e) instanceof NullPointerException);
                     }
                 }
             }